Support for easily including full-width images
Adds a \fullwidthimage command, which takes two parameters: the image
to be included, and whether the image should be placed at the top or
the bottom of the page. Furthermore, users can now optionally pass an
alternative representation of the list of authors  to \author, and the
definition of the \maketitle command now includes an extra newline
character.