Graphics path and extension detection for \includepdf
\includepdf now mimics the behavior of \includegraphics, and will also
attempt to look for files in ./Images/Screen/ and ./Images/Print, if
those folders exist.

Unlike \includegraphics (which remains unmodified), \includepdf
currently does not allow one to refer to the full name (i.e. including
the file extension) of image files. This may or may not change in
future versions.

Finally, the order of preference for image types has been changed to
.jpg, .pdf, .png, the reason being that PDF files are more likely to
contain vector images which can be resized without reduction in image
quality.