\includepdf works with full paths to images again
Since commit 5121aa0, \includepdf no longer allowed users to insert
images by passing a full path (i.e. with the image’s file extension).

Not allowing image extensions to be hardcoded made it a lot easier to
“hot-swap” image files for other (CMYK and compressed) versions, but
unfortunately it also broke some stuff, and somewhat runs counter to
the principle that the bonaparticle class should never get in the way
of the user.