Remove all references to Frutiger
As per announcement of commit 6f1e570, all references to the Frutiger
font will now be removed from the “develop” and “master” branches.

The main reason for this change is that the Frutiger font is not part
of the bonaparticle class per se, and thus should not be included in
this repository. Other reasons for omitting the Frutiger files include
licensing and support.