Contain each included document in a group
By forcing each included document to be contained within a group, it’s
much harder for an included document to adversely affect any of the
documents that come after it.

Some examples of things that would happen: larger font sizes,
inconsistent margins, and wrapfigures which create large amounts of
whitespace on subsequent pages.