Make infoboxes the right size in a multicols environment
The problem was that the minipage had a size of \textwidth, i.e. the
total width that text occupies on a page, instead of \linewidth, i.e.
the width of each individual line. In multicols environments, using the
\textwidth makes an infobox take the width of two columns instead of one.

We might want to look for other cases of \textwidth that should be
\linewidth.